fix: erros in code logic fixed

+++ ExpenseTracking/Models/ExpenseEntry.cs
@@ -4,7 +4,7 @@ namespace ExpenseTracking.models
 {
     public class ExpenseEntry : FinancialEntry
     {
-        public ExpenseEntry(int id, DateOnly date, int value, string description, string category) : base(id,date,value,description,category)
+        public ExpenseEntry(int id, DateOnly date, float value, string description, string category) : base(id,date,value,description,category)
         {}
     }
 }
