+++ ExpenseTracking/Models/FinancialEntry.cs
@@ -5,11 +5,11 @@ namespace ExpenseTracking.models
     {
         public int _id;
         public DateOnly Date{ get; set; }
-        public int Value { get; set; }
+        public float Value { get; set; }
         public string Description { get; set; }
         public string Category { get; set; }
 
-        public FinancialEntry(int id, DateOnly date, int value, string description, string category)
+        public FinancialEntry(int id, DateOnly date, float value, string description, string category)
         {
             _id = id;
             Date = date;
