+++ ExpenseTracking/Models/RevenueEntry.cs
@@ -4,7 +4,7 @@ namespace ExpenseTracking.models
 {
     public class RevenueEntry : FinancialEntry
     {
-        public RevenueEntry(int id,DateOnly date, int value, string description, string category) : base(id,date, value, description, category)
+        public RevenueEntry(int id,DateOnly date, float value, string description, string category) : base(id,date, value, description, category)
         { }
     }
 }
